refactor: get_language function moved into utils
