+++ src/utils/get_language.ts
@@ -1,0 +1,29 @@
+// Determine language based on file extension
+const getLanguage = (filePath: string) => {
+  const extension = filePath.split(".").pop()?.toLowerCase() || "";
+  const languageMap: Record<string, string> = {
+    js: "javascript",
+    jsx: "javascript",
+    ts: "typescript",
+    tsx: "typescript",
+    html: "html",
+    css: "css",
+    json: "json",
+    md: "markdown",
+    py: "python",
+    java: "java",
+    c: "c",
+    cpp: "cpp",
+    cs: "csharp",
+    go: "go",
+    rs: "rust",
+    rb: "ruby",
+    php: "php",
+    swift: "swift",
+    kt: "kotlin",
+    // Add more as needed
+  };
+
+  return languageMap[extension] || "plaintext";
+};
+export { getLanguage };
